+++ packages/core/acme-client/test/10-verify.spec.js
@@ -17,6 +17,10 @@ describe('verify', () => {
     const testHttp01Challenge = { type: 'http-01', status: 'pending', token: uuid() };
     const testHttp01Key = uuid();
 
+    const testHttps01Authz = { identifier: { type: 'dns', value: `${uuid()}.${domainName}` } };
+    const testHttps01Challenge = { type: 'http-01', status: 'pending', token: uuid() };
+    const testHttps01Key = uuid();
+
     const testDns01Authz = { identifier: { type: 'dns', value: `${uuid()}.${domainName}` } };
     const testDns01Challenge = { type: 'dns-01', status: 'pending', token: uuid() };
     const testDns01Key = uuid();
@@ -69,6 +73,27 @@ describe('verify', () => {
 
         it('should verify challenge with trailing newline', async () => {
             const resp = await verify['http-01'](testHttp01Authz, testHttp01Challenge, testHttp01Key);
+            assert.isTrue(resp);
+        });
+    });
+
+
+    /**
+     * https-01
+     */
+
+    describe('https-01', () => {
+        it('should reject challenge', async () => {
+            await assert.isRejected(verify['http-01'](testHttps01Authz, testHttps01Challenge, testHttps01Key));
+        });
+
+        it('should mock challenge response', async () => {
+            const resp = await cts.addHttps01ChallengeResponse(testHttps01Challenge.token, testHttps01Key, testHttps01Authz.identifier.value);
+            assert.isTrue(resp);
+        });
+
+        it('should verify challenge', async () => {
+            const resp = await verify['http-01'](testHttps01Authz, testHttps01Challenge, testHttps01Key);
             assert.isTrue(resp);
         });
     });